+++ CHANGELOG.md
@@ -68,4 +68,4 @@
 
 # v6.0.0
 
-See the [v6.0.0 release page](https://invertase.io/oss/react-native-firebase/releases/v6.0.0).
+See the [v6.0.0 release page](https://rnfirebase.io/releases/v6.0.0).